cfx_rpc_cfx_impl

Module helpers

Re-exports

pub use block_provider::build_block;

pub use block_provider::build_header;

pub use poll_filter::limit_logs;

pub use poll_filter::PollFilter;

pub use poll_filter::SyncPollFilter;

pub use poll_filter::MAX_BLOCK_HISTORY_SIZE;

pub use poll_manager::PollManager;

Modules

block_provider

poll_filter

poll_manager

subscribers

A map of subscribers.

Structs

EpochQueue

The goal of EpochQueue is to keep a distance from the tip of the ledger. This way, we can ensure that the epoch being processed has already been executed (deferred execution) and we can reduce the number of chain reorgs that we need to notify subscribers about.

Constants

MAX_FEE_HISTORY_CACHE_BLOCK_COUNT